(PC-37544)[API] chore: move dms commands
dms/commands.py is not sourced by `install_commands` but
subscription/commands.py is
